pyapi/requirements.txt
fastapi[standard]>=0.115.8,<0.116.0
sqlalchemy>=2.0.37,<3.0.0
databases>=0.9.0,<0.10.0
requests>=2.32.3,<3.0.0
uvicorn>=0.34.0,<0.35.0
loguru==0.7.3
apscheduler==3.11.0
azure-cognitiveservices-speech==1.42.0
twilio==9.4.4
openai==1.61.0
python-dotenv==1.0.1
passlib==1.7.4
